+++ cucumber-expressions/java/src/main/java/io/cucumber/cucumberexpressions/CucumberExpression.java
@@ -11,10 +11,10 @@ public class CucumberExpression implements Expression {
     private static final Pattern OPTIONAL_PATTERN = Pattern.compile("\\(([^\\)]+)\\)");
 
     private final Pattern pattern;
-    private final List<Transform<?>> transforms = new ArrayList<>();
+    private final List<Parameter<?>> parameters = new ArrayList<>();
     private final String expression;
 
-    public CucumberExpression(final String expression, List<? extends Type> types, TransformLookup transformLookup) {
+    public CucumberExpression(final String expression, List<? extends Type> types, ParameterRegistry parameterRegistry) {
         this.expression = expression;
         String expressionWithOptionalGroups = OPTIONAL_PATTERN.matcher(expression).replaceAll("(?:$1)?");
         Matcher matcher = PARAMETER_PATTERN.matcher(expressionWithOptionalGroups);
@@ -27,25 +27,25 @@ public class CucumberExpression implements Expression {
             String parameterName = matcher.group(1);
             String typeName = matcher.group(3);
 
-            Transform<?> transform = null;
+            Parameter<?> parameter = null;
             if (type != null) {
-                transform = transformLookup.lookupByType(type);
+                parameter = parameterRegistry.lookupByType(type);
             }
-            if (transform == null && typeName != null) {
-                transform = transformLookup.lookupByTypeName(typeName, false);
+            if (parameter == null && typeName != null) {
+                parameter = parameterRegistry.lookupByTypeName(typeName, false);
             }
-            if (transform == null) {
-                transform = transformLookup.lookupByTypeName(parameterName, true);
+            if (parameter == null) {
+                parameter = parameterRegistry.lookupByTypeName(parameterName, true);
             }
-            if (transform == null && type != null && type instanceof Class) {
-                transform = new ClassTransform<>((Class) type);
+            if (parameter == null && type != null && type instanceof Class) {
+                parameter = new ClassParameter<>((Class) type);
             }
-            if (transform == null) {
-                transform = new ConstructorTransform<>(String.class);
+            if (parameter == null) {
+                parameter = new ConstructorParameter<>(String.class);
             }
-            transforms.add(transform);
+            parameters.add(parameter);
 
-            matcher.appendReplacement(regexp, Matcher.quoteReplacement(getCaptureGroupRegexp(transform.getCaptureGroupRegexps())));
+            matcher.appendReplacement(regexp, Matcher.quoteReplacement(getCaptureGroupRegexp(parameter.getCaptureGroupRegexps())));
         }
         matcher.appendTail(regexp);
         regexp.append("$");
@@ -73,7 +73,7 @@ public class CucumberExpression implements Expression {
 
     @Override
     public List<Argument> match(String text) {
-        return ArgumentMatcher.matchArguments(pattern, text, transforms);
+        return ArgumentBuilder.buildArguments(pattern, text, parameters);
     }
 
     @Override