+++ cucumber-expressions/java/src/main/java/io/cucumber/cucumberexpressions/ExpressionFactory.java
@@ -18,33 +18,33 @@ public class ExpressionFactory {
     private static final Pattern SCRIPT_STYLE_REGEXP = Pattern.compile("^/(.*)/$");
     private static final Pattern PARENS = Pattern.compile("\\(([^\\)]+)\\)");
     private static final Pattern ALPHA = Pattern.compile("[a-zA-Z]+");
-    private final TransformLookup transformLookup;
+    private final ParameterRegistry parameterRegistry;
 
-    public ExpressionFactory(TransformLookup transformLookup) {
-        this.transformLookup = transformLookup;
+    public ExpressionFactory(ParameterRegistry parameterRegistry) {
+        this.parameterRegistry = parameterRegistry;
     }
 
     public Expression createExpression(String expressionString, List<Type> types) {
         Matcher m = BEGIN_ANCHOR.matcher(expressionString);
         if (m.find()) {
-            return new RegularExpression(Pattern.compile(expressionString), types, transformLookup);
+            return new RegularExpression(Pattern.compile(expressionString), types, parameterRegistry);
         }
         m = END_ANCHOR.matcher(expressionString);
         if (m.find()) {
-            return new RegularExpression(Pattern.compile(expressionString), types, transformLookup);
+            return new RegularExpression(Pattern.compile(expressionString), types, parameterRegistry);
         }
         m = SCRIPT_STYLE_REGEXP.matcher(expressionString);
         if (m.find()) {
-            return new RegularExpression(Pattern.compile(m.group(1)), types, transformLookup);
+            return new RegularExpression(Pattern.compile(m.group(1)), types, parameterRegistry);
         }
         m = PARENS.matcher(expressionString);
         if (m.find()) {
             String insideParens = m.group(1);
             if (ALPHA.matcher(insideParens).lookingAt()) {
-                return new CucumberExpression(expressionString, types, transformLookup);
+                return new CucumberExpression(expressionString, types, parameterRegistry);
             }
-            return new RegularExpression(Pattern.compile(expressionString), types, transformLookup);
+            return new RegularExpression(Pattern.compile(expressionString), types, parameterRegistry);
         }
-        return new CucumberExpression(expressionString, types, transformLookup);
+        return new CucumberExpression(expressionString, types, parameterRegistry);
     }
 }