+++ cucumber-expressions/javascript/src/parameter.js
@@ -1,9 +1,9 @@
-class Transform {
-  constructor(typeName, constructorFunction, captureGroupRegexps, transformer) {
+class Parameter {
+  constructor(typeName, constructorFunction, captureGroupRegexps, transform) {
     this._typeName = typeName
     this._constructorFunction = constructorFunction
     this._captureGroupRegexps = stringArray(captureGroupRegexps)
-    this._transformer = transformer
+    this._transform = transform
   }
 
   get typeName() {
@@ -19,7 +19,7 @@ class Transform {
   }
 
   transform(string) {
-    return this._transformer(string)
+    return this._transform(string)
   }
 }
 
@@ -28,4 +28,4 @@ function stringArray(captureGroupRegexps) {
   return array.map(r => typeof r == 'string' ? r : r.source)
 }
 
-module.exports = Transform
+module.exports = Parameter