+++ cucumber-expressions/ruby/lib/cucumber/cucumber_expressions/parameter.rb
@@ -1,13 +1,13 @@
 module Cucumber
   module CucumberExpressions
-    class Transform
+    class Parameter
       attr_reader :type_name, :type, :capture_group_regexps
 
-      # Create a new Transform
+      # Create a new Parameter
       #
       # @param type_name [Array] array of class or type name to use in {arg:type_name}
       # @param capture_group_regexps [Array] list of regexps for capture groups.
-      # @param transformer lambda that transforms a String to another type
+      # @param transformer lambda that transforms a String to (possibly) another type
       #
       def initialize(type_name, type, capture_group_regexps, transformer)
         @type_name, @type, @transformer = type_name, type, transformer