+++ cucumber-expressions/ruby/spec/cucumber/cucumber_expressions/regular_expression_spec.rb
@@ -1,47 +1,47 @@
 require 'cucumber/cucumber_expressions/regular_expression'
-require 'cucumber/cucumber_expressions/transform_lookup'
+require 'cucumber/cucumber_expressions/parameter_registry'
 
 module Cucumber
   module CucumberExpressions
     describe RegularExpression do
       it "documents match arguments" do
-        transform_lookup = TransformLookup.new
+        parameter_registry = ParameterRegistry.new
 
         ### [capture-match-arguments]
         expr = /I have (\d+) cukes? in my (\w*) now/
         types = ['int', nil]
-        expression = RegularExpression.new(expr, types, transform_lookup)
+        expression = RegularExpression.new(expr, types, parameter_registry)
         args = expression.match("I have 7 cukes in my belly now")
         expect( args[0].transformed_value ).to eq(7)
         expect( args[1].transformed_value ).to eq("belly")
         ### [capture-match-arguments]
       end
 
-      it "transforms to string by default" do
+      it "does no transform by default" do
         expect( match(/(\d\d)/, "22") ).to eq(["22"])
       end
 
-      it "transforms integer to double using explicit type name" do
+      it "transforms int to float by explicit type name" do
         expect( match(/(.*)/, "22", ['float']) ).to eq([22.0])
       end
 
-      it "transforms integer to double using explicit type" do
+      it "transforms int to float by explicit function" do
         expect( match(/(.*)/, "22", [Float]) ).to eq([22.0])
       end
 
-      it "transforms to int using capture group pattern" do
+      it "transforms int by parameter pattern" do
         expect( match(/(-?\d+)/, "22") ).to eq([22])
       end
 
-      it "transforms to int by alternate capture group pattern" do
+      it "transforms int by alternate parameter pattern" do
         expect( match(/(\d+)/, "22") ).to eq([22])
       end
 
-      it "transforms double without integer value" do
+      it "transforms float without integer part" do
         expect( match(/(.*)/, ".22", ['float']) ).to eq([0.22])
       end
 
-      it "transforms double with sign" do
+      it "transforms float with sign" do
         expect( match(/(.*)/, "-1.22", ['float']) ).to eq([-1.22])
       end
 
@@ -57,11 +57,11 @@ module Cucumber
 
       it "exposes source" do
         expr = /I have (\d+) cukes? in my (\+) now/
-        expect(RegularExpression.new(expr, [], TransformLookup.new).source).to eq(expr)
+        expect(RegularExpression.new(expr, [], ParameterRegistry.new).source).to eq(expr)
       end
 
       def match(expression, text, types = [])
-        regular_expression = RegularExpression.new(expression, types, TransformLookup.new)
+        regular_expression = RegularExpression.new(expression, types, ParameterRegistry.new)
         arguments = regular_expression.match(text)
         return nil if arguments.nil?
         arguments.map { |arg| arg.transformed_value }